+++ src/App.js
@@ -2,14 +2,14 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import MainLayout from "./components/layouts/MainLayout";
 import HomePage from "./components/features/homePage/HomePage";
 import SearchPage from "./components/features/searchPage/SearchPage";
-import RestaurantDetailPage from "./components/features/detailPage/RestaurantDetailPage";
+import RestaurantDetailPage from "./components/features/detailPage/restaurantDetailPage/RestaurantDetailPage";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { Provider } from "react-redux";
 import FoodSearchPage from "./components/features/foodSearchPage/FoodSearchPage";
 import WishlistPage from "./components/features/wishlistPage/WishlistPage";
-import FestivalDetailPage from "./components/features/detailPage/FestivalDetailPage";
+import FestivalDetailPage from "./components/features/detailPage/festivalDetailPage/FestivalDetailPage";
 import ReservationListPage from "./components/features/reservationListPage/ReservationListPage";
-import FoodDetailPage from "./components/features/foodDetailPage/FoodDetailPage";
+import FoodDetailPage from "./components/features/detailPage/foodDetailPage/FoodDetailPage";
 
 import LoginPage from "./components/features/formPages/loginPage/LoginPage";
 import RegisterPage from "./components/features/formPages/registerPage/RegisterPage";
