+++ src/components/features/detailPage/festivalDetailPage/FestivalDetailPage.jsx
@@ -1,9 +1,9 @@
 import { useParams } from "react-router-dom";
 import { useQuery } from "react-query";
-import { getFestivalById } from "../../../apis/detail";
+import { getFestivalById } from "../../../../apis/detail";
 import FestivalDetailTemplate from "./FestivalDetailTemplate";
-import ErrorPage from "../ErrorPage/ErrorPage";
-import LoadingPage from "../loadingPage/LoadingPage";
+import ErrorPage from "../../ErrorPage/ErrorPage";
+import LoadingPage from "../../loadingPage/LoadingPage";
 
 const FestivalDetailPage = () => {
   const params = useParams().id;
