+++ src/components/features/detailPage/festivalDetailPage/FestivalDetailTemplate.jsx
@@ -1,41 +1,39 @@
-import PageTitleBar from "../../molecules/PageTitleBar";
+import PageTitleBar from "../../../molecules/PageTitleBar";
+import SectionTitle from "../../../atoms/SectionTitle";
+import InfoElement from "../atoms/InfoElement";
+import { comma } from "../../../../utils/convert";
+import ReviewCards from "../../../molecules/cards/ReviewCards";
+import ButtonAllReviews from "../atoms/ButtonAllReviews";
 import { useState } from "react";
-import ReviewCards from "../../molecules/cards/ReviewCards";
-import SectionTitle from "../../atoms/SectionTitle";
-import HorizontalListSection from "../carousel/HorizontalListSection";
-import MenuCard from "../../molecules/cards/MenuCard";
-import ButtonAllReviews from "./atoms/ButtonAllReviews";
 import { useQuery } from "react-query";
-import { getReviewByIdAndType } from "../../../apis/review";
-import BottomPopModal from "../../atoms/Modals/BottomPopModal";
-import InfoElement from "./atoms/InfoElement";
-import { getCalenderByIdAndType } from "../../../apis/detail";
-import Calendar from "../calendar/Calendar";
-import Photo from "../../atoms/Photo";
-import Button from "../../atoms/Button";
-import TimeDropdown from "../../molecules/TimeDropdown";
-import CardTitle from "../../atoms/CardTitle";
-import { reserveRestaurant } from "../../../apis/reservation";
+import { getReviewByIdAndType } from "../../../../apis/review";
+import BottomPopModal from "../../../atoms/Modals/BottomPopModal";
+import Calendar from "../../calendar/Calendar";
+import { getCalenderByIdAndType } from "../../../../apis/detail";
+import Button from "../../../atoms/Button";
+import Photo from "../../../atoms/Photo";
+import TimeDropdown from "../../../molecules/TimeDropdown";
+import CardTitle from "../../../atoms/CardTitle";
+import { reserveFestival } from "../../../../apis/reservation";
 import { useNavigate } from "react-router-dom";
-import Article from "../../organisms/Article";
+import Article from "../../../organisms/Article";
 
-const RestaurantDetailTemplate = ({ restaurant }) => {
+const FestivalDetailTemplate = ({ festival }) => {
   const [isActiveReview, setIsActiveReview] = useState(false);
   const [isActiveCalender, setIsActiveCalender] = useState(false);
   const [selectedDate, setSelectedDate] = useState(null);
-  const [requestMessage, setRequestMessage] = useState("");
   const [selectedTime, setSelectedTime] = useState("Time To Visit");
   const [selectedPeople, setSelectedPeople] = useState(0);
 
-  const { data } = useQuery(`restaurant/review/${restaurant.id}`, () =>
-    getReviewByIdAndType(restaurant.id, "restaurant"),
-  );
-
   const navigate = useNavigate();
 
+  const { data } = useQuery(`festival/review/${festival.id}`, () =>
+    getReviewByIdAndType(festival.id, "festival"),
+  );
+
   const { data: operatingInfo } = useQuery(
-    `restaurant/unavailableDays/${restaurant.id}`,
-    () => getCalenderByIdAndType(restaurant.id, "restaurant"),
+    `festival/unavailableDays/${festival.id}`,
+    () => getCalenderByIdAndType(festival.id, "festival"),
   );
 
   const onReserve = async () => {
@@ -47,12 +45,11 @@ const RestaurantDetailTemplate = ({ restaurant }) => {
       alert("Please select date to visit");
       return;
     }
-    const response = await reserveRestaurant(
-      restaurant.id,
+    const response = await reserveFestival(
+      festival.id,
       selectedDate,
       selectedTime,
       selectedPeople,
-      requestMessage,
     );
     if (response.success) {
       alert("Reservation success");
@@ -63,8 +60,8 @@ const RestaurantDetailTemplate = ({ restaurant }) => {
   };
 
   return (
-    <div className={"restaurant-detail-template w-full"}>
-      <PageTitleBar name={restaurant.name} />
+    <div className={"festival-detail-template w-full"}>
+      <PageTitleBar name={festival.name} />
       {(isActiveReview || isActiveCalender) && (
         <BottomPopModal
           onClose={() => {
@@ -74,15 +71,13 @@ const RestaurantDetailTemplate = ({ restaurant }) => {
         >
           {isActiveReview && <ReviewCards reviews={data.reviews} />}
           {isActiveCalender && (
-            <div
-              className={"calendar-wrapper flex flex-col justify-center px-2"}
-            >
+            <div className={"calendar-wrapper flex flex-col justify-center"}>
               <Calendar
                 selectedDate={selectedDate}
                 setSelectedDate={setSelectedDate}
                 unavailableDays={operatingInfo.holiday}
               />
-              <div className={"time-select-form flex flex-col py-2 text-lg"}>
+              <div className={"time-select-form flex flex-col p-2 text-lg"}>
                 <CardTitle title={"Visit Time"} />
                 <div className={"dropdown-wrapper"}>
                   <TimeDropdown
@@ -113,68 +108,49 @@ const RestaurantDetailTemplate = ({ restaurant }) => {
                     }}
                   />
                 </div>
+                <Button
+                  as="button"
+                  onClick={onReserve}
+                  variant="link"
+                  className="rounded-button-[tripKoOrange] mt-4 flex h-12 w-full items-center justify-center rounded-full bg-tripKoOrange text-white"
+                >
+                  Reservation
+                </Button>
               </div>
-              <div className={"request-message-form"}>
-                <CardTitle title={"Request Message"} />
-                <textarea
-                  className={
-                    "request-message-input h-20 w-full rounded-md border-2 border-gray-300 p-2"
-                  }
-                  placeholder={"Please enter your request message"}
-                  value={requestMessage}
-                  onChange={(e) => setRequestMessage(e.target.value)}
-                />
-              </div>
-              <Button
-                as="button"
-                onClick={onReserve}
-                variant="link"
-                className="rounded-button-[tripKoOrange] my-2 h-12 w-full rounded-full bg-tripKoOrange text-white"
-              >
-                Reserve
-              </Button>
             </div>
           )}
         </BottomPopModal>
       )}
       <div
         className={
-          "restaurant-image-wrapper width-flex-layout fixed top-0 w-full "
+          "festival-image-wrapper width-flex-layout fixed top-0 w-full "
         }
       >
         <Photo
-          src={restaurant.mainImage}
-          alt={restaurant.name}
-          className={"min-h-[30rem] w-full"}
+          src={festival.mainImage}
+          alt={festival.name}
+          className={"min-h-[50rem] w-full"}
           extendable={true}
         />
       </div>
       <div
         className={
-          "restaurant-detail-content relative mt-[30rem] bg-white pb-[8rem]"
+          "festival-detail-content relative mt-[50rem] bg-white pb-[8rem]"
         }
       >
-        <SectionTitle title={"Menu"} />
-        <HorizontalListSection>
-          {restaurant.menu.map((menu, index) => (
-            <MenuCard menu={menu} key={index} />
-          ))}
-        </HorizontalListSection>
-        <SectionTitle title={"Information"} />
-        {restaurant.content.map((content) => (
+        <SectionTitle title={"Information"}/>
+        {festival.contents.map((content) => (
           <Article
-            key={content.page}
-            description={content.description}
-            images={content.image}
+              key={content.page}
+              description={content.description}
+              images={content.image}
           />
         ))}
         <div className={"information-card grid gap-2 px-4 py-2 md:grid-cols-2"}>
-          <InfoElement title={"Address"} value={restaurant.address} />
-          <InfoElement title={"Contact"} value={restaurant.contactInfo} />
-          <InfoElement title={"Operating Hours"} value={restaurant.open} />
-          <InfoElement title={"Break Time"} value={restaurant.breakTime} />
+          <InfoElement title={"Address"} value={festival.address} />
+          <InfoElement title={"Period"} value={festival.period} />
+          <InfoElement title={"Price"} value={`₩${comma(festival.price)}`} />
         </div>
-        <SectionTitle title={"Photo"} />
         <SectionTitle title={"Reviews"} />
         {data && <ReviewCards reviews={data.reviews.slice(0, 2)} />}
         <ButtonAllReviews onClick={() => setIsActiveReview(true)} />
@@ -196,4 +172,4 @@ const RestaurantDetailTemplate = ({ restaurant }) => {
   );
 };
 
-export default RestaurantDetailTemplate;
+export default FestivalDetailTemplate;
