+++ src/components/features/detailPage/restaurantDetailPage/RestaurantDetailPage.jsx
@@ -1,9 +1,9 @@
 import { useParams } from "react-router-dom";
 import { useQuery } from "react-query";
-import { getRestaurantById } from "../../../apis/detail";
+import { getRestaurantById } from "../../../../apis/detail";
 import RestaurantDetailTemplate from "./RestaurantDetailTemplate";
-import LoadingPage from "../loadingPage/LoadingPage";
-import ErrorPage from "../ErrorPage/ErrorPage";
+import LoadingPage from "../../loadingPage/LoadingPage";
+import ErrorPage from "../../ErrorPage/ErrorPage";
 
 const RestaurantDetailPage = () => {
   const params = useParams().id;
