+++ src/components/features/detailPage/restaurantDetailPage/RestaurantDetailTemplate.jsx
@@ -1,39 +1,41 @@
-import PageTitleBar from "../../molecules/PageTitleBar";
-import SectionTitle from "../../atoms/SectionTitle";
-import InfoElement from "./atoms/InfoElement";
-import { comma } from "../../../utils/convert";
-import ReviewCards from "../../molecules/cards/ReviewCards";
-import ButtonAllReviews from "./atoms/ButtonAllReviews";
+import PageTitleBar from "../../../molecules/PageTitleBar";
 import { useState } from "react";
+import ReviewCards from "../../../molecules/cards/ReviewCards";
+import SectionTitle from "../../../atoms/SectionTitle";
+import HorizontalListSection from "../../carousel/HorizontalListSection";
+import MenuCard from "../../../molecules/cards/MenuCard";
+import ButtonAllReviews from "../atoms/ButtonAllReviews";
 import { useQuery } from "react-query";
-import { getReviewByIdAndType } from "../../../apis/review";
-import BottomPopModal from "../../atoms/Modals/BottomPopModal";
-import Calendar from "../calendar/Calendar";
-import { getCalenderByIdAndType } from "../../../apis/detail";
-import Button from "../../atoms/Button";
-import Photo from "../../atoms/Photo";
-import TimeDropdown from "../../molecules/TimeDropdown";
-import CardTitle from "../../atoms/CardTitle";
-import { reserveFestival } from "../../../apis/reservation";
+import { getReviewByIdAndType } from "../../../../apis/review";
+import BottomPopModal from "../../../atoms/Modals/BottomPopModal";
+import InfoElement from "../atoms/InfoElement";
+import { getCalenderByIdAndType } from "../../../../apis/detail";
+import Calendar from "../../calendar/Calendar";
+import Photo from "../../../atoms/Photo";
+import Button from "../../../atoms/Button";
+import TimeDropdown from "../../../molecules/TimeDropdown";
+import CardTitle from "../../../atoms/CardTitle";
+import { reserveRestaurant } from "../../../../apis/reservation";
 import { useNavigate } from "react-router-dom";
-import Article from "../../organisms/Article";
+import Article from "../../../organisms/Article";
 
-const FestivalDetailTemplate = ({ festival }) => {
+const RestaurantDetailTemplate = ({ restaurant }) => {
   const [isActiveReview, setIsActiveReview] = useState(false);
   const [isActiveCalender, setIsActiveCalender] = useState(false);
   const [selectedDate, setSelectedDate] = useState(null);
+  const [requestMessage, setRequestMessage] = useState("");
   const [selectedTime, setSelectedTime] = useState("Time To Visit");
   const [selectedPeople, setSelectedPeople] = useState(0);
 
-  const navigate = useNavigate();
-
-  const { data } = useQuery(`festival/review/${festival.id}`, () =>
-    getReviewByIdAndType(festival.id, "festival"),
+  const { data } = useQuery(`restaurant/review/${restaurant.id}`, () =>
+    getReviewByIdAndType(restaurant.id, "restaurant"),
   );
 
+  const navigate = useNavigate();
+
   const { data: operatingInfo } = useQuery(
-    `festival/unavailableDays/${festival.id}`,
-    () => getCalenderByIdAndType(festival.id, "festival"),
+    `restaurant/unavailableDays/${restaurant.id}`,
+    () => getCalenderByIdAndType(restaurant.id, "restaurant"),
   );
 
   const onReserve = async () => {
@@ -45,11 +47,12 @@ const FestivalDetailTemplate = ({ festival }) => {
       alert("Please select date to visit");
       return;
     }
-    const response = await reserveFestival(
-      festival.id,
+    const response = await reserveRestaurant(
+      restaurant.id,
       selectedDate,
       selectedTime,
       selectedPeople,
+      requestMessage,
     );
     if (response.success) {
       alert("Reservation success");
@@ -60,8 +63,8 @@ const FestivalDetailTemplate = ({ festival }) => {
   };
 
   return (
-    <div className={"festival-detail-template w-full"}>
-      <PageTitleBar name={festival.name} />
+    <div className={"restaurant-detail-template w-full"}>
+      <PageTitleBar name={restaurant.name} />
       {(isActiveReview || isActiveCalender) && (
         <BottomPopModal
           onClose={() => {
@@ -71,13 +74,15 @@ const FestivalDetailTemplate = ({ festival }) => {
         >
           {isActiveReview && <ReviewCards reviews={data.reviews} />}
           {isActiveCalender && (
-            <div className={"calendar-wrapper flex flex-col justify-center"}>
+            <div
+              className={"calendar-wrapper flex flex-col justify-center px-2"}
+            >
               <Calendar
                 selectedDate={selectedDate}
                 setSelectedDate={setSelectedDate}
                 unavailableDays={operatingInfo.holiday}
               />
-              <div className={"time-select-form flex flex-col p-2 text-lg"}>
+              <div className={"time-select-form flex flex-col py-2 text-lg"}>
                 <CardTitle title={"Visit Time"} />
                 <div className={"dropdown-wrapper"}>
                   <TimeDropdown
@@ -108,48 +113,66 @@ const FestivalDetailTemplate = ({ festival }) => {
                     }}
                   />
                 </div>
-                <Button
-                  as="button"
-                  onClick={onReserve}
-                  variant="link"
-                  className="rounded-button-[tripKoOrange] mt-4 flex h-12 w-full items-center justify-center rounded-full bg-tripKoOrange text-white"
-                >
-                  Reservation
-                </Button>
               </div>
+              <div className={"request-message-form"}>
+                <CardTitle title={"Request Message"} />
+                <textarea
+                  className={
+                    "request-message-input h-20 w-full rounded-md border-2 border-gray-300 p-2"
+                  }
+                  placeholder={"Please enter your request message"}
+                  value={requestMessage}
+                  onChange={(e) => setRequestMessage(e.target.value)}
+                />
+              </div>
+              <Button
+                as="button"
+                onClick={onReserve}
+                variant="link"
+                className="rounded-button-[tripKoOrange] my-2 h-12 w-full rounded-full bg-tripKoOrange text-white"
+              >
+                Reserve
+              </Button>
             </div>
           )}
         </BottomPopModal>
       )}
       <div
         className={
-          "festival-image-wrapper width-flex-layout fixed top-0 w-full "
+          "restaurant-image-wrapper width-flex-layout fixed top-0 w-full "
         }
       >
         <Photo
-          src={festival.mainImage}
-          alt={festival.name}
-          className={"min-h-[50rem] w-full"}
+          src={restaurant.mainImage}
+          alt={restaurant.name}
+          className={"min-h-[30rem] w-full"}
           extendable={true}
         />
       </div>
       <div
         className={
-          "festival-detail-content relative mt-[50rem] bg-white pb-[8rem]"
+          "restaurant-detail-content relative mt-[30rem] bg-white pb-[8rem]"
         }
       >
-        <SectionTitle title={"Information"}/>
-        {festival.contents.map((content) => (
+        <SectionTitle title={"Menu"} />
+        <HorizontalListSection>
+          {restaurant.menu.map((menu, index) => (
+            <MenuCard menu={menu} key={index} />
+          ))}
+        </HorizontalListSection>
+        <SectionTitle title={"Information"} />
+        {restaurant.contents.map((content) => (
           <Article
-              key={content.page}
-              description={content.description}
-              images={content.image}
+            key={content.page}
+            description={content.description}
+            images={content.image}
           />
         ))}
         <div className={"information-card grid gap-2 px-4 py-2 md:grid-cols-2"}>
-          <InfoElement title={"Address"} value={festival.address} />
-          <InfoElement title={"Period"} value={festival.period} />
-          <InfoElement title={"Price"} value={`₩${comma(festival.price)}`} />
+          <InfoElement title={"Address"} value={restaurant.address} />
+          <InfoElement title={"Contact"} value={restaurant.contactInfo} />
+          <InfoElement title={"Operating Hours"} value={restaurant.open} />
+          <InfoElement title={"Break Time"} value={restaurant.breakTime} />
         </div>
         <SectionTitle title={"Reviews"} />
         {data && <ReviewCards reviews={data.reviews.slice(0, 2)} />}
@@ -172,4 +195,4 @@ const FestivalDetailTemplate = ({ festival }) => {
   );
 };
 
-export default FestivalDetailTemplate;
+export default RestaurantDetailTemplate;
